add generator testing to Value to Json
